backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
websockets==13.0
pydantic==2.9.2
apscheduler==3.10.4
python-dotenv==1.0.1
sqlalchemy[asyncio]==2.0.35
aiomysql==0.2.0
pymysql==1.1.1
